Refactor portal sections preparation for portal service updates
