+++ workspace/apps/pidp/src/app/features/portal/enums/alert-code.enum.ts
@@ -1,0 +1,4 @@
+export enum AlertCode {
+  TRANSIENT_ERROR = 1,
+  PLR_BAD_STANDING,
+}
