+++ workspace/apps/pidp/src/app/features/portal/enums/status-code.enum.ts
@@ -1,0 +1,6 @@
+export enum StatusCode {
+  AVAILABLE = 1,
+  COMPLETED,
+  NOT_AVAILABLE,
+  ERROR,
+}
