+++ workspace/apps/pidp/src/app/features/portal/models/profile-status-alert.model.ts
@@ -1,0 +1,4 @@
+export interface ProfileStatusAlert {
+  heading: string;
+  content: string;
+}
